requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
streamlit
langchain_ollama